requirements.txt
numpy
tqdm
matplotlib
h5py
pyyaml
scipy